Enable automatic documentation of serializers
This patch adds support for automatic discovery of serializer fields and
their properties. The idea is to simplify creating documentation.

This is not totally universal, there are many cases where the serializer
will not be documented, in other places some docstrings are required.

The information about serializers is cached, so that each serializer is
explored at most twice (once with read-only fields, once without).